refactor: thermo-nuclear review follow-ups (#590-#599)
Consolidate Atlas reducers and Hermes typing, split tearsheet charts into
submodules, tighten score/CI path config, and harden digichat BFF caches.

Part of #579. Fixes #590, #591, #592, #593, #594, #595, #596, #597, #598, #599.
