+++ .github/workflows/ci.yml
@@ -40,6 +40,7 @@ jobs:
         id: filter
         with:
           filters: |
+            # CI_PATH_FILTERS_START — generated from scripts/ci_paths.yaml
             digibase:
               - 'digibase/**'
               - 'tests/db/**'
@@ -125,6 +126,7 @@ jobs:
               - 'agents.yml'
               - 'agents/**'
               - '.github/workflows/ci.yml'
+              - 'scripts/ci_paths.yaml'
             compose:
               - 'docker-compose.yml'
               - '.env.example'
@@ -134,6 +136,7 @@ jobs:
               - '**/pyproject.toml'
               - 'pip-audit-ignore.txt'
               - '.github/workflows/pip-audit.yml'
+            # CI_PATH_FILTERS_END
 
   # ── Per-component test suites (gated by paths) ───────────────────────────────
   digibase:
