+++ digiquant/src/digiquant/hermes/state.py
@@ -7,13 +7,10 @@ until digest-only extraction lands (epic #471).
 from __future__ import annotations
 
 from digiquant.atlas.state import (
-    AnalystRowPayload,
     AtlasResearchState,
-    DebateTickerState,
     Phase7DigestPayload,
     RebalancePayload,
     RiskDebatePayload,
-    SpecialistAxisPayload,
 )
 
 # Alias — see module docstring. New code should import ``HermesState`` from
@@ -21,12 +18,9 @@ from digiquant.atlas.state import (
 HermesState = AtlasResearchState
 
 __all__ = [
-    "AnalystRowPayload",
     "AtlasResearchState",
-    "DebateTickerState",
     "HermesState",
     "Phase7DigestPayload",
     "RebalancePayload",
     "RiskDebatePayload",
-    "SpecialistAxisPayload",
 ]
