+++ scripts/check_pandas_boundary.sh
@@ -8,7 +8,9 @@ cd "$ROOT"
 ALLOWLIST=(
   "digiquant/src/digiquant/nautilus_runner.py"
   "digiquant/src/digiquant/tearsheet.py"
-  "digiquant/src/digiquant/tearsheet_charts.py"
+  "digiquant/src/digiquant/charts/returns.py"
+  "digiquant/src/digiquant/charts/drawdown.py"
+  "digiquant/src/digiquant/charts/equity.py"
   "digiquant/scripts/atlas/preload-history.py"
   "digiquant/src/digiquant/strategies/bollinger_mr.py"
   "digiquant/src/digiquant/strategies/macd_trend.py"
