+++ scripts/ci_paths.yaml
@@ -1,0 +1,114 @@
+# Single source of truth for CI path filters (dorny/paths-filter in ci.yml).
+# Regenerate embedded block: python3 scripts/generate_ci_path_filters.py
+# Drift check: python3 scripts/generate_ci_path_filters.py --check
+
+digibase:
+  - digibase/**
+  - tests/db/**
+  - .github/workflows/digibase-test.yml
+
+digikey:
+  - digikey/**
+  - tests/dk/**
+  - .github/workflows/digikey-test.yml
+
+digismith:
+  - digismith/**
+  - tests/dsm/**
+  - .github/workflows/digismith-test.yml
+
+digiclaw:
+  - digiclaw/**
+  - tests/dc/**
+  - .github/workflows/digiclaw-test.yml
+
+digiquant:
+  - digiquant/**
+  - tests/dq/**
+  - .github/workflows/digiquant-test.yml
+
+digisearch:
+  - digisearch/**
+  - tests/ds/**
+  - .github/workflows/digisearch-test.yml
+
+digigraph:
+  - digigraph/**
+  - tests/dg/**
+  - .github/workflows/digigraph-test.yml
+
+digichat:
+  - frontend/digichat/**
+  - frontend/design/**
+  - package.json
+  - package-lock.json
+  - .github/workflows/digichat-test.yml
+
+olympus:
+  - frontend/olympus/**
+  - frontend/design/**
+  - package.json
+  - package-lock.json
+  - .github/workflows/olympus-test.yml
+
+score:
+  - digibase/**
+  - digiclaw/**
+  - digigraph/**
+  - digiquant/**
+  - digisearch/**
+  - digismith/**
+  - digikey/**
+  - frontend/**
+  - scripts/**
+  - tests/**
+  - agents.yml
+  - agents/**
+  - .github/workflows/score-pr.yml
+
+e2e_contract:
+  - tests/test_e2e.py
+  - tests/test_e2e_contract.py
+  - docker-compose.yml
+  - digikey/**
+  - .github/workflows/e2e.yml
+
+nautilus_smoke:
+  - digiquant/**
+  - tests/dq/test_nautilus_runner.py
+  - .github/workflows/nautilus-smoke.yml
+
+atlas_graph:
+  - digiquant/src/digiquant/atlas/**
+  - digiquant/scripts/atlas/**
+  - digiquant/supabase/**
+  - tests/dq/atlas/**
+  - digiquant/src/digiquant/hermes/**
+  - tests/dq/hermes/**
+  - .github/workflows/atlas-graph-ci.yml
+  - .github/workflows/atlas-*.yml
+
+ruff_and_scripts:
+  - digibase/**
+  - digiclaw/**
+  - digigraph/**
+  - digiquant/**
+  - digisearch/**
+  - digismith/**
+  - tests/**
+  - scripts/**
+  - agents.yml
+  - agents/**
+  - .github/workflows/ci.yml
+  - scripts/ci_paths.yaml
+
+compose:
+  - docker-compose.yml
+  - .env.example
+  - "**/Dockerfile"
+  - docker/**
+
+pip_audit:
+  - "**/pyproject.toml"
+  - pip-audit-ignore.txt
+  - .github/workflows/pip-audit.yml
